Extract prop_compose test module
So that the `proptest::Strategy`s can be reused without needing to
depend on the `payout_curve` submodule.
